# api/openapi.yaml
openapi: 3.1.0
info:
  title: dimsum control API
  version: 1.0.0
  description: |
    All generation and job IDs are decimal strings. Content revisions and transaction IDs
    are opaque SHA256 strings. No unconfigured statistics dependency returns fabricated data.
    Browser-session mutations require Origin and X-CSRF-Token obtained at login.
    Opaque bearer tokens grant full administrator access without browser Origin/CSRF
    credentials. Tokens have a separate lifetime from browser sessions and remain
    valid until explicitly revoked. At most 32 tokens may be active; plaintext is
    returned only at creation. Collections use YAML
    schema field names. PATCH paths are relative to the collection (including index),
    or absolute configuration paths for settings. Unsupported comment-preserving edits
    fail explicitly. Successful saves include saved and active configuration identities.
    History windows are half-open [from,to), positive, and at most 366 days. Timestamps
    use UTC with at most microsecond precision. Defaults end at the current UTC minute
    and start one hour earlier; when only to is provided, from defaults to to minus
    one hour. All endpoints share these defaults. updated_at is the response computation
    time, not a promise of durable observation coverage; complete carries that meaning.
    Reuse explicit range values across requests for a fixed dashboard window.
...
security:
  - session: []
  - bearer: []
paths:
  /api/v1/dhcp:
    x-mcp-tools:
      get: {name: get_dhcp, description: Inspect saved DHCP settings and configuration revision before editing}
      patch: {name: update_dhcp, description: Apply revision-checked DHCP scalar edits then inspect get_dhcp_status for applied generation; after a timeout inspect before retrying}
    get:
      responses:
        '200':
          description: Saved settings; disabled incomplete configuration is editable without opening sockets or lease storage
          content:
            application/json:
              schema:
                type: object
                required: [status, config]
                properties:
                  status: {$ref: '#/components/schemas/Activation'}
                  config: {$ref: '#/components/schemas/DHCPSettings'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      description: Relative scalar paths only. Disable before changing interface, server_ip, subnet or local_domain while serving. Reservations use their own collection. A saved/active configuration generation does not mean DHCP has applied it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DHCPSettingsMutation'}
      responses:
        '200': {$ref: '#/components/responses/Activation'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/queries:
    x-mcp-tools:
      get: {name: list_queries, description: List a bounded page of retained DNS queries}
    get:
      description: |
        Descending timestamp and ID keyset page with a fixed insertion ceiling.
        Cursor binds normalized filters and window. With a cursor, omitted from/to
        reuse its original window; repeat the same filters on every page. Malformed
        cursors and changed filters/range return 400 bad_request. Unknown or repeated
        parameters are rejected. No offset pagination or substring-name search.
      parameters:
        - {$ref: '#/components/parameters/From'}
        - {$ref: '#/components/parameters/To'}
        - {$ref: '#/components/parameters/Limit'}
        - {name: cursor, in: query, schema: {type: string, maxLength: 2048}}
        - {name: name, in: query, description: 'Exact name; accepts canonical DNS decimal escapes from displayed names, or dot for root.', schema: {type: string}}
        - {name: client, in: query, description: Literal observed IPv4 or IPv6 address without zone, schema: {type: string}}
        - {name: outcome, in: query, schema: {$ref: '#/components/schemas/Outcome'}}
        - {name: qtype, in: query, description: 'DNS type mnemonic, decimal 1–65535, or TYPE<number>', schema: {type: string}}
        - {name: boot_id, in: query, description: 'Exact process boot identity; required with rule_id or upstream_id', schema: {type: string, minLength: 1, maxLength: 128}}
        - {name: generation, in: query, description: 'Canonical uint32 decimal configuration generation; required with rule_id or upstream_id', schema: {type: string, pattern: '^(0|[1-9][0-9]*)$'}}
        - {name: rule_id, in: query, description: 'Canonical uint32 decimal rule number, scoped by boot_id and generation', schema: {type: string, pattern: '^(0|[1-9][0-9]*)$'}}
        - {name: upstream_id, in: query, description: 'Canonical uint16 decimal upstream number, scoped by boot_id and generation because endpoints can change on reload', schema: {type: string, pattern: '^(0|[1-9][0-9]*)$'}}
        - {name: source_id, in: query, description: 'Exact archived source identity, at most 512 UTF-8 bytes. Legacy or oversized identities are unavailable, never inferred from descriptions.', schema: {type: string, minLength: 1, maxLength: 512}}
      responses:
        '200':
          description: Retained query page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HistoryQueries'}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/rules/test:
    x-mcp-tools:
      post: {name: explain_domain, description: Explain a domain decision with the active policy matcher without performing a DNS lookup, readOnly: true}
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string}
                qtype: {type: string, description: Name-policy matcher is QTYPE independent}
                generation: {type: string, pattern: '^[0-9]+$'}
      responses:
        '200': {description: 'name, normalized, generation, decision from active policy matcher; not a DNS lookup'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/config/transactions:
    x-mcp-tools:
      post: {name: stage_configuration, description: Validate and stage revision-checked settings edits without activating them}
    post:
      <<: *patch
      description: Validate/stage settings edits without activation
      responses:
        '200': {description: 'id (opaque content hash) and revision; commit explicitly'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  parameters:
    From:
      name: from
      in: query
      description: Inclusive UTC RFC3339 timestamp; default is to minus one hour.
      schema: {type: string, format: date-time}
    To:
      name: to
      in: query
      description: Exclusive UTC RFC3339 timestamp; default is current UTC minute.
      schema: {type: string, format: date-time}
    Limit:
      name: limit
      in: query
      schema: {type: integer, default: 100, minimum: 1, maximum: 200}
  securitySchemes:
    session: {type: apiKey, in: cookie, name: dimsum_session, description: Browser-session authentication; mutations also require Origin and X-CSRF-Token}
    bearer: {type: http, scheme: bearer, description: Opaque full-administrator token with independent lifetime until explicit revocation; checked on every HTTP request including MCP requests; no browser Origin/CSRF credentials required}
  schemas:
...
    Decimal:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
      description: Unsigned integer encoded as a decimal string, including values beyond JavaScript exact integers.
    HistoryRange:
      type: object
      required: [from, to]
      properties:
        from: {type: string, format: date-time}
        to: {type: string, format: date-time}
    Outcome:
      type: string
      enum: [local, blocked, cache, stale, forwarded, error, rejected]
...
    QueryResponse:
      type: object
      description: Bounded summary of the reply sent to the client. Omitted on a query when response data was not recorded; no new DNS lookup is performed.
      required: [records, truncated]
      properties:
        truncated: {type: boolean, description: Some records or values were omitted by capture/summary limits; independent of the DNS TC flag}
        records:
          type: array
          maxItems: 16
          items:
            type: object
            required: [name, type, value, ttl, section]
            properties:
              name: {type: string, description: Byte-safe record owner; dot denotes root}
              type: {type: string, description: DNS type mnemonic or TYPE<number>}
              value: {type: string, maxLength: 1024, description: Record presentation; unsupported types use RFC 3597-style hexadecimal data}
              ttl: {type: integer, minimum: 0, maximum: 4294967295, description: TTL in seconds as returned to this client at query time, not a live countdown}
              section: {type: string, enum: [answer, authority, additional]}
    HistoryQuery:
      type: object
      required: [id, boot_id, sequence, time, client, client_name, client_name_source, client_name_fresh, name, qtype, qtype_code, qclass, outcome, rcode, duration_us, generation, rule_id, upstream_id, flags]
      properties:
        id: {$ref: '#/components/schemas/Decimal'}
        boot_id: {type: string, description: Boot scope for generation and rule identifiers}
        sequence: {$ref: '#/components/schemas/Decimal'}
        time: {type: string, format: date-time}
        client: {type: string, description: Observed address, always present independently of its friendly name}
        client_name: {type: string, description: Current cached friendly name; empty when unknown, not a historical identity claim}
        client_name_source: {type: string}
        client_name_expires: {type: string, format: date-time, description: Expiry of the derived client name; omitted for authoritative configured names.}
        client_name_fresh: {type: boolean}
        client_device: {$ref: '#/components/schemas/DeviceEnrichment'}
        name: {type: string, description: Lowercase byte-safe DNS presentation with decimal escapes; root is empty}
        qtype: {type: string, description: DNS type mnemonic or TYPE<number>}
        qtype_code: {type: integer, minimum: 0, maximum: 65535}
        qclass: {type: integer, minimum: 0, maximum: 65535}
        outcome: {$ref: '#/components/schemas/Outcome'}
        rcode: {type: integer, minimum: 0, maximum: 65535}
        duration_us: {$ref: '#/components/schemas/Decimal'}
        generation: {$ref: '#/components/schemas/Decimal'}
        rule_id: {$ref: '#/components/schemas/Decimal'}
        source_id: {type: string, description: 'Exact archived source identity; empty when unavailable (including legacy history and identities exceeding 512 bytes)'}
        upstream_id: {$ref: '#/components/schemas/Decimal'}
        flags: {type: integer, minimum: 0, maximum: 4294967295, description: 'Bits 0 TCP, 1 coalesced, 2 fallback used, 3 truncated, 4 response-policy block'}
        response: {$ref: '#/components/schemas/QueryResponse'}
    HistoryQueries:
      type: object
      required: [items, complete, range, updated_at]
      properties:
        items: {type: array, maxItems: 200, items: {$ref: '#/components/schemas/HistoryQuery'}}
        next_cursor: {type: string, description: Omitted on the last page; opaque base64url value}
        complete: {type: boolean, description: Observation/retention completeness, not whether there are more pages}
        range: {$ref: '#/components/schemas/HistoryRange'}
        updated_at: {type: string, format: date-time}
...
    DeviceEvidence:
      type: object
      required: [source, updated, expires]
      properties:
        source: {type: string}
        hostname: {type: string, maxLength: 1024}
        service_type: {type: string, maxLength: 1024}
        label: {type: string, maxLength: 256}
        model: {type: string, maxLength: 256}
        manufacturer: {type: string, maxLength: 256}
        device_type: {type: string, maxLength: 256}
        updated: {type: string, format: date-time}
        expires: {type: string, format: date-time}
    DeviceEnrichment:
      type: object
      description: Current derived device presentation; explicit client names always take priority over discovered names. Not an authenticated physical identity.
      required: [category, reason, inferred, fresh, evidence]
      properties:
        category: {type: string, enum: [unknown, phone, tablet, laptop, desktop, tv, speaker, printer, camera, lighting, appliance, server, console]}
        reason: {type: string}
        inferred: {type: boolean}
        fresh: {type: boolean}
        hostname: {type: string}
        model: {type: string}
        manufacturer: {type: string}
        evidence: {type: array, maxItems: 16, items: {$ref: '#/components/schemas/DeviceEvidence'}}
        dns_guess: {$ref: '#/components/schemas/DNSGuess'}
    DNSGuess:
      type: object
      description: Lowest-priority device-name guess based on recent DNS questions, not verified product identity.
      required: [rule, reason, domains, expires]
      properties:
        rule: {type: string}
        reason: {type: string}
        expires: {type: string, format: date-time}
        domains:
          type: array
          maxItems: 16
          items:
            type: object
            required: [domain, first_seen, last_seen, queries]
            properties:
              domain: {type: string}
              first_seen: {type: string, format: date-time}
              last_seen: {type: string, format: date-time}
              queries: {type: string, pattern: '^[0-9]+$'}
...
    Activation:
      type: object
      required: [saved_revision, active_revision, active_generation, pending, recovered, restart_required, sources]
      properties:
        saved_revision: {type: string}
        active_revision: {type: string}
        active_generation: {type: string, pattern: '^[0-9]+$'}
        pending: {type: boolean}
        recovered: {type: boolean}
        restart_required: {type: boolean}
        error: {type: string}
        sources: {type: array, items: {type: object}}
        dashboard_hosts:
          description: Unaccepted exact local A/AAAA names pointing to this admin listener; returned after saving records. Approval is a separate revision-checked records PATCH.
          type: array
          items:
            type: object
            required: [name, host, url]
            properties:
              name: {type: string}
              host: {type: string}
              url: {type: string, format: uri}
...
    DHCPSettings:
      type: object
      required: [enabled, interface, server_ip, subnet, gateway, range_start, range_end, lease_seconds, local_domain, max_leases, reservations]
      properties:
        enabled: {type: boolean, default: false}
        interface: {type: string, maxLength: 15}
        server_ip: {type: string, description: Static IPv4 server and advertised DNS address; empty allowed while disabled}
        subnet: {type: string, description: Canonical IPv4 subnet /0 through /30; empty allowed while disabled}
        gateway: {type: string, description: IPv4 router outside dynamic pool and different from server IP}
        range_start: {type: string, description: Inclusive IPv4 pool start; maximum span 65536 addresses}
        range_end: {type: string, description: Inclusive IPv4 pool end}
        lease_seconds: {type: integer, minimum: 0, maximum: 604800, description: Advertised lease duration in seconds; 60..604800 when supplied, 0 allowed only while disabled}
        local_domain: {type: string, maxLength: 189, description: DNS domain outside .local; empty allowed while disabled}
        max_leases: {type: integer, minimum: 0, maximum: 4096, description: Lease-state capacity; 0 selects default 1024}
        reservations: {type: array, maxItems: 4096, items: {$ref: '#/components/schemas/DHCPReservation'}}
    DHCPReservation:
      type: object
      additionalProperties: false
      required: [id, address]
      description: Exactly one nonempty mac or client_id. Address must be usable in the configured subnet; IDs, identities, addresses and supplied hostnames must be unique.
      properties:
        id: {type: string, minLength: 1, maxLength: 63, pattern: '^[A-Za-z0-9]([A-Za-z0-9-]*[A-Za-z0-9])?$'}
        mac: {type: string, maxLength: 17, description: Unicast Ethernet MAC; empty means absent}
        client_id: {type: string, maxLength: 510, pattern: '^([0-9a-fA-F]{2})*$', description: Opaque client ID bytes encoded as hex; empty means absent}
        address: {type: string, format: ipv4}
        hostname: {type: string, maxLength: 63}
    DHCPSettingsMutation:
      type: object
      additionalProperties: false
      required: [revision, edits]
      properties:
        revision: {type: string, minLength: 1}
        edits:
          type: array
          minItems: 1
          maxItems: 16
          items:
            type: object
            additionalProperties: false
            required: [path, value]
            properties:
              path: {type: array, minItems: 1, maxItems: 1, items: {type: string, enum: [enabled, interface, server_ip, subnet, gateway, range_start, range_end, lease_seconds, local_domain, max_leases]}}
              value: {type: [string, integer, boolean], description: Field-appropriate scalar; lease_seconds and max_leases are integers, enabled is boolean}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, field_errors]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}
            active_generation: {type: string, pattern: '^[0-9]+$'}
            field_errors: {type: array, items: {type: object}}
  responses:
    Activation:
      description: Validated and activated configuration
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Activation'}
    Error:
      description: '400 malformed/bad_request, 401 auth, 403 CSRF/Host/Origin, 404 missing or expired detail, 409 revision/generation conflict, 422 validation, 429 capacity, 503 unavailable'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}